browser-use-test/requirements.txt
browser-use>=0.1.40,<0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
